[endstop] Move EndstopActuatorBase from actuator/ to endstop/ component
- Move endstop_actuator.h/.cpp from actuator/ to endstop/
- Change namespace from esphome::actuator to esphome::endstop
- Move ENDSTOP_ACTUATOR_SCHEMA and apply_endstop_actuator_config from
  actuator/__init__.py to endstop/__init__.py
- Update endstop/cover.py to import from local package instead of actuator
- Update endstop/endstop_cover.h to include local header and use
  unqualified EndstopActuatorBase (same namespace)
- Strip endstop helpers from actuator/__init__.py; actuator now only
  contains shared base infrastructure (ActuatorBase, ActuatorCallBase,
  IActuator, ActuatorOperation, build_apply_lambda_action)
